mathmodel-Backend/requirements.txt
biopython>=1.78
numpy>=1.18.5
scipy>=1.5.0
py-expression-eval>=0.3.14
matplotlib>=3.2.2
scikit-learn>=0.23.1
flask-cors
